+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.4] - 2024-0
+### Added
+- Import/Export All config
+- Backup to local Download path
+- Empty button CSS
+
 ### Changed
 - Form with labels instead of table for `Add/Edit` server